+++ src/i8c/compiler/driver.py
@@ -24,6 +24,7 @@ from __future__ import unicode_literals
 from .. import cmdline
 from ..compat import fprint
 from . import blocks
+from . import commands
 from . import emitter
 from . import externals
 from . import I8CError
@@ -179,7 +180,7 @@ class CommandLine(object):
 def setup_input(args):
     process = infile = None
     if args.with_cpp:
-        command = ["gcc", "-E", "-x", "c", "-D__INFINITY__"] + args.cpp_args
+        command = commands.I8C_CPP + ["-D__INFINITY__"] + args.cpp_args
         process = subprocess.Popen(command, stdout=subprocess.PIPE)
         infile = process.stdout
     elif args.infiles in ([], ["-"]):
@@ -205,7 +206,7 @@ def guess_outfile(args):
 
 def setup_output(args):
     if args.with_asm:
-        command = ["gcc", "-x", "assembler-with-cpp"] + args.asm_args + ["-"]
+        command = commands.I8C_AS + args.asm_args + ["-"]
         if args.outfile is None and "-c" in args.asm_args:
             command.extend(("-o", guess_outfile(args)))
         process = subprocess.Popen(command, stdin=subprocess.PIPE)